Formatting + Handling server errors
ValueError and KeyError are literally always connected to server
issues. Handling them as such might stop users opening up new issues
asking for them to be fixed.

I’m quite new to python, so if there is a way to print traceback as
well as error, go ahead and fix it :)