convert ods file to release_config
I created a script which allows the user to insert values into an ods (excel/calc) file and transfer it into the json file.
I would like to use a new folder for it but since I'm new to github, I don't know how to create one. 
The example release file is in ods format and needs to be added.