Increases fort movement efficiency
Makes it so distance tracking is done after every fort, instead of just
at the beginning. This should make walking times between pokestops much
less in certain cases. (For example, on the old loop, if fort A was
2.5mi North, fort B was 2.7 mi North, and fort C was 2.6mi South, the
program would go A->C->B, which is insanely inefficient)

Also, the self.position in __init__.py now stays up-to-date with the
actual current position. This was needed to allow for the more real-time
location checking.