Windows installation instructions refit (#3898)
* Delete PokemonGo-Bot-Repair.bat

* Update PokemonGo-Bot-Install.bat

* Update PokemonGo-Bot-Start.bat

Removed an update procedure that was making the map not to work.

* Update installation.md

Modified Windows installation instructions to use the batch files.

* Update PokemonGo-Bot-Install.bat

Added links for the configuration instructions. Still missing links for the encrypt files. Need help on it.

* Update manual_installation.md

Recreated the installations instructions for Windows.

* Update installation.md

Removed the need to download PyYAML and modified the instructions to download only the required encrypt files.

* Update manual_installation.md

Removed the requirement to download PyYAML as the installer already downloads it during git pull. Also removed the requirement to download all encrypt files.

* Update PokemonGo-Bot-Install.bat

* Update PokemonGo-Bot-Install.bat

* Update PokemonGo-Bot-Start.bat

Added a pause in the end so the promp wont close in the end if `pokecli.py` goes to error.

* Update PokemonGo-Bot-Install.bat

Added a procedure to deal with the installation path issue. Also, added a command to install protobuff automatically.

* Update installation.md

Removed the requirement to install protobuff as it will be installed automatically by the install batch.

* Update PokemonGo-Bot-Install.bat

* Update manual_installation.md

Removed the requirement for protoc and added command line to download it.

* Update PokemonGo-Bot-Install.bat

* Update PokemonGo-Bot-Start.bat

Updated in order to work anywhere as long its inside the windows_bat folder

* Update PokemonGo-Bot-Install.bat

* Update installation.md

* Update PokemonGo-Bot-Start.bat

* Update PokemonGo-Bot-Install.bat

* Update CONTRIBUTORS.md

* Update PokemonGo-Bot-Start.bat

* Update PokemonGo-Bot-Start.bat

* Update PokemonGo-Bot-Start.bat

* Update PokemonGo-Bot-Install.bat

* Revert "Update CONTRIBUTORS.md"

This reverts commit 856a0fd5e4fc76c8d3b6d8b8da0000571923d74b.

* Revert "Delete PokemonGo-Bot-Repair.bat"

This reverts commit a459a8661e0c402354dbe7df4ce4dfef5618a52c.

* Revert "Update PokemonGo-Bot-Start.bat"

This reverts commit 510570d3868535fc47d3addbd0a25794bd7f843f.

* Revert "Update PokemonGo-Bot-Install.bat"

This reverts commit 9db27989a9df93b65974bc8a8fc8323f68a14e83.

# Conflicts:
#	windows_bat/PokemonGo-Bot-Install.bat

* Complete refit of all batches

We can now choose an install location and all required software is
automatically installed.

* Added contributors

Added new contributors

* Revert "Added contributors"

This reverts commit a1d364c924581256de14af0e01ec21c71a91c722.