Documenting docker-compose with Tor (#4161)
* Send bot through Tor

All of the bot's API requests go through tor

* links privoxy

you gotta link if you want to reference

* keep privoxy up too

* Docker documentation update

Including example command for using the alternate tor proxy command. Additionally changed line 114 to down rather than rm so that the containers are stopped before they are removed in the case that the containers were launched in detached mode.