Use Incense worker added
I saw a PR a while back (PR #3755) by bmorrise but it was closed, not
sure why (except for some comments)
I'm new at python programming, so I hope I've adjusted the code in this
version well enough.

This will make the bot use the Incense in the bag when it's available.
No pause, no interval, when you have it, use it.