CampFort - Check pokeballs before camping (#5177)
* CampFort - Check pokeballs before camping

I noticed that the bot doesn't move when camping at a lure, but had run
out of pokeballs. This will first check IF we have any pokeballs left.
If we have no pokeballs left, abandon the lure (to move to a fort to get
more pokeballs) when we got some pokeballs again, will move back to the
lure.

* Moved check down, added announce once

As suggested, moved the check a bit down
Solved the message spam with a switch to announce or not

* Adjusted code based on advise by Anakin5

Now the worker will set a time to walk away from the lures.
After the time has passed, the worker will check again if we have some
Pokeballs. If we have enough, we will search for a lure again (the old
one could be finished)

* For get the old location