Remove error for empty reply in sniper feed
When starting up the bot, if the sniper feed has 0 items in it, it disables sniping for the rest of the session. This might make sense on the surface for something like a RocketMap feed, where there should always be at least something, but when using a custom feed with only amazing spawns, it's often empty. This isn't a problem - the sniper loops through all items in the feed, which in this case is zero, then moves on. Actual problems with the feed (404, invalid JSON, etc) will trigger a separate error. 

I've tested this with a valid feed with items, a valid feed with zero items, and invalid feeds - all work as expected.