Pokemon Hunter - Mobster and Trash Hunter
In this version of the Pokemon Hunter, we can select if we want to hunt
down family members of VIPs (Like a mobster) and if we want to hunt for
certain "trash" Pokemon.

Mobster
When setting "target_family_of_vip"to true, the Hunter will also mark
family of the VIPs you set as a valid target.
If you have "lock_on" enabled, and want to be sure you catch every last
family member of the VIPs, you can enable "treat_family_of_vip_as_vip".
This effectively means that the whole family of a VIP needs to be hunted
down as VIP Pokemons.

Trash Hunting
Trash hunting will trigger when a set amount of space is left in the
bag. If triggered, the Hunter will start hunting down Caterpies, Weedles
and Pidgeys. These only take 12 candy to evolve, so are useful for mass
evolving.

If a VIP Pokemon is detected nearby while hunting for trash, the Hunter
will change the target to the VIP (and locking onto it if enabled)