This eventhandler update writes also to a json file in web folder
With this change, the interesting events are also written to a json file in web folder. This allows the web-ui to display the event messages on screen. With this the web-ui can now see in detail what the bot is actually doing by refreshing the json file.

With         self.MaxEventlog = 50
you can limit the amount of lines to save on disk. This to avoid the json file gets way to large to deal with!